Assignment-3/requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
Werkzeug==2.3.7
